+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 
 - refactored `render_mqtt()` to `render_template()` in `main.py`. This better reflects that this is for rendering any Jinja variable type
+- JinjaVariable types only update on the sign if the rendered template has actually changed due to a data change
 
 ## Version 2.8